feat(desktop): list recent chats in the conversation rail
Scoping the rail to its route (#651) took the chat list out of a
conversation, so switching chats meant going back to home first — the
longest path for the navigation done most. Brightwave keeps a recent-
conversations list in the rail that is visible inside a conversation,
so this ports that shape rather than adding a separate switcher.

The recent list (collapsible header, rows with rename/delete, per-row
attention markers) is extracted from HomeSidebar into a shared
RecentChatsSection and mounted in ChatSidebar too, with the open chat
marked as current. The back link keeps its elsewhere-needs-attention
marker for chats beyond the list's cut.

Closes #655
