+++ crates/openwave-desktop/ui/src/sidebar/HomeSidebar.tsx
@@ -1,17 +1,12 @@
-import { useState } from "react";
 import { useNavigate, useRouterState, useSearch } from "@tanstack/react-router";
-import { History, LayoutGrid, MessagesSquare } from "lucide-react";
+import { LayoutGrid, MessagesSquare } from "lucide-react";
 
 import { useApp } from "@/AppContext";
-import { useChatAttention } from "@/ChatAttention";
 import { useChatListStore } from "@/ChatListStore";
 import { NewChatButton } from "./NewChatButton";
-import { RecentChatRow } from "./RecentChatRow";
-import { SidebarButton, SidebarSectionTitle, useSidebarWidth } from "./primitives";
+import { RecentChatsSection } from "./RecentChatsSection";
+import { SidebarButton, SidebarSectionTitle } from "./primitives";
 import { SidebarFrame } from "./SidebarFrame";
-
-/** How many conversations the rail shows before deferring to the All chats table. */
-const RECENT_CHAT_LIMIT = 8;
 
 /**
  * The rail outside any conversation: starting one, and returning to one.
@@ -26,22 +21,12 @@ const RECENT_CHAT_LIMIT = 8;
  */
 export function HomeSidebar() {
   const navigate = useNavigate();
-  const { newChat, deleteChat, startRename, commitRename, cancelRename, refreshChats } = useApp();
-  const chats = useChatListStore((state) => state.chats);
+  const { newChat, refreshChats } = useApp();
   const chatsError = useChatListStore((state) => state.chatsError);
   const creatingChat = useChatListStore((state) => state.creatingChat);
   const deletingChatId = useChatListStore((state) => state.deletingChatId);
-  const renamingChatId = useChatListStore((state) => state.renamingChatId);
-  const renameChatDraft = useChatListStore((state) => state.renameChatDraft);
-  const savingTitle = useChatListStore((state) => state.savingTitle);
-  const setRenameDraft = useChatListStore((state) => state.setRenameDraft);
-  const chatIdsWithPendingPrompts = useChatAttention(
-    (state) => state.chatIdsWithPendingPrompts,
-  );
-  const isCompact = useSidebarWidth() === "compact";
   const pathname = useRouterState({ select: (state) => state.location.pathname });
   const search = useSearch({ strict: false }) as { left?: string; right?: string };
-  const [recentCollapsed, setRecentCollapsed] = useState(false);
 
   // The Apps library lives on home; the entry is "current" when either slot
   // holds it, list or detail alike.
@@ -49,9 +34,6 @@ export function HomeSidebar() {
     segment === "apps" || segment?.startsWith("apps.") === true;
   const appsOpen =
     pathname === "/" && (holdsApps(search.left) || holdsApps(search.right));
-
-  const recentChats = chats.slice(0, RECENT_CHAT_LIMIT);
-  const showRecent = !recentCollapsed && !isCompact && recentChats.length > 0;
 
   return (
     <SidebarFrame>
@@ -63,38 +45,7 @@ export function HomeSidebar() {
 
       <SidebarSectionTitle className="mt-4">Chats</SidebarSectionTitle>
       <div className="flex flex-col gap-0.5">
-        <SidebarButton
-          aria-expanded={!recentCollapsed}
-          onClick={() => setRecentCollapsed((collapsed) => !collapsed)}
-        >
-          <History />
-          <span>Recent</span>
-        </SidebarButton>
-        {showRecent && (
-          <div
-            className="ml-4 flex flex-col gap-0.5 border-l border-border pl-2"
-            aria-label="Recent chats"
-          >
-            {recentChats.map((chat) => (
-              <RecentChatRow
-                key={chat.id}
-                chat={chat}
-                active={false}
-                needsAttention={chatIdsWithPendingPrompts.has(chat.id)}
-                renaming={renamingChatId === chat.id}
-                renameDraft={renameChatDraft}
-                savingTitle={savingTitle}
-                mutating={deletingChatId !== null || creatingChat}
-                onRenameDraftChange={setRenameDraft}
-                onOpen={() => void navigate({ to: "/c/$chatId", params: { chatId: chat.id } })}
-                onStartRename={() => startRename(chat)}
-                onCommitRename={() => commitRename(chat)}
-                onCancelRename={cancelRename}
-                onDelete={() => deleteChat(chat)}
-              />
-            ))}
-          </div>
-        )}
+        <RecentChatsSection />
         <SidebarButton
           aria-current={pathname === "/chats" ? "page" : undefined}
           data-active={pathname === "/chats" || undefined}
